control flow + functional guards
