+++ src/EDAWebClient/src/app/app.component.html
@@ -1,8 +1,9 @@
-<div *ngIf="isReady" class="app-main-wrapper">
-  <app-navbar></app-navbar>
-  <div class="root-container">
-    <router-outlet></router-outlet>
+@if (isReady) {
+  <div class="app-main-wrapper">
+    <app-navbar></app-navbar>
+    <div class="root-container">
+      <router-outlet></router-outlet>
+    </div>
+    <app-snack-bar></app-snack-bar>
   </div>
-
-  <app-snack-bar></app-snack-bar>
-</div>
+}
