+++ src/EDAWebClient/src/app/components/home/home.component.html
@@ -7,49 +7,53 @@
             placeholder="Select process"
             (selectionChange)="onSelectProcess($event.value)"
             [(value)]="selectedProcessId"
-          >
-            <mat-option
-              value="{{ option.ProcessID }}"
-              *ngFor="let option of processes"
-              >{{ option.ProcessName }}</mat-option
             >
-          </mat-select>
-        </mat-form-field>
-      </div>
-      <div
-        class="description-area"
-        title="{{ longDescription || shortDescription }}"
-      >
-        <span class="label">Short description:</span>
-        <span class="text" *ngIf="shortDescription">{{
-          shortDescription
-        }}</span>
-        <span
-          class="empty-description"
-          *ngIf="!shortDescription && !longDescription"
-          class="empty-description f-14"
-          >Empty</span
-        >
-      </div>
-    </div>
+            @for (option of processes; track option) {
+              <mat-option
+                value="{{ option.ProcessID }}"
+                >{{ option.ProcessName }}</mat-option
+                >
+              }
+            </mat-select>
+          </mat-form-field>
+        </div>
+        <div
+          class="description-area"
+          title="{{ longDescription || shortDescription }}"
+          >
+          <span class="label">Short description:</span>
+          @if (shortDescription) {
+            <span class="text">{{
+              shortDescription
+            }}</span>
+          }
+          @if (!shortDescription && !longDescription) {
+            <span
+              class="empty-description"
+              class="empty-description f-14"
+              >Empty</span
+              >
+            }
+          </div>
+        </div>
 
-    <app-upload></app-upload>
+        <app-upload></app-upload>
 
-    <div class="actions">
-      <button
-        [disabled]="!isProcessSelected"
-        mat-raised-button
-        color="primary"
-        (click)="retrieveResults()"
-      >
-        Retrieve results
-      </button>
+        <div class="actions">
+          <button
+            [disabled]="!isProcessSelected"
+            mat-raised-button
+            color="primary"
+            (click)="retrieveResults()"
+            >
+            Retrieve results
+          </button>
+        </div>
+      </div>
+      <aside class="sidebar-content">
+        <app-settings></app-settings>
+      </aside>
+      <footer class="footer-content">
+        <app-logs></app-logs>
+      </footer>
     </div>
-  </div>
-  <aside class="sidebar-content">
-    <app-settings></app-settings>
-  </aside>
-  <footer class="footer-content">
-    <app-logs></app-logs>
-  </footer>
-</div>
