+++ src/EDAWebClient/src/app/components/logs/logs.component.html
@@ -1,11 +1,15 @@
 <div class="logs-wrapper">
   <div class="logs-container">
-    <div *ngFor="let log of logsList" class="log-entry">
-      <div [innerHTML]="log | urlCustomPipe"></div>
-    </div>
+    @for (log of logsList; track log) {
+      <div class="log-entry">
+        <div [innerHTML]="log | urlCustomPipe"></div>
+      </div>
+    }
 
-    <div *ngIf="logsList?.length === 0" class="empty-logs">
-      No logs to display.
-    </div>
+    @if (logsList?.length === 0) {
+      <div class="empty-logs">
+        No logs to display.
+      </div>
+    }
   </div>
 </div>
