+++ src/EDAWebClient/src/app/components/settings-mappings/settings-mappings.component.html
@@ -6,10 +6,12 @@
         [disabled]="!isProcessSelected"
         (selectionChange)="onSelectDocumentType($event.value)"
         [(value)]="selectedDocumentTypeId"
-      >
-        <mat-option [value]="option.ID" *ngFor="let option of documentTypes">
-          {{ option.Name }}
-        </mat-option>
+        >
+        @for (option of documentTypes; track option) {
+          <mat-option [value]="option.ID">
+            {{ option.Name }}
+          </mat-option>
+        }
       </mat-select>
     </mat-form-field>
 
@@ -18,50 +20,49 @@
       mat-raised-button
       color="warn"
       (click)="onReset()"
-    >
+      >
       Reset
     </button>
   </div>
 
-  <div class="mappings-container" *ngIf="isProcessSelected">
-    <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-      <ng-container matColumnDef="source">
-        <th mat-header-cell *matHeaderCellDef>Source</th>
-        <td
-          mat-cell
-          *matCellDef="let element"
-          class="col-m-w ellipsis"
-          title="{{ element.Name }}"
-        >
-          {{ element.Name }}
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="destination">
-        <th mat-header-cell *matHeaderCellDef>Destination</th>
-        <td
-          mat-cell
-          *matCellDef="let element"
-          class="col-m-w ellipsis"
-          title="{{ element.Destination }}"
-        >
-          {{ element.Destination }}
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="actions">
-        <th mat-header-cell *matHeaderCellDef class="text-right">Actions</th>
-        <td mat-cell *matCellDef="let element" class="text-right">
-          <mat-icon
-            class="action-icon fill-primary"
-            svgIcon="edit_ico"
-            (click)="onEdit(element)"
-          ></mat-icon>
-        </td>
-      </ng-container>
-
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-    </table>
-  </div>
+  @if (isProcessSelected) {
+    <div class="mappings-container">
+      <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
+        <ng-container matColumnDef="source">
+          <th mat-header-cell *matHeaderCellDef>Source</th>
+          <td
+            mat-cell
+            *matCellDef="let element"
+            class="col-m-w ellipsis"
+            title="{{ element.Name }}"
+            >
+            {{ element.Name }}
+          </td>
+        </ng-container>
+        <ng-container matColumnDef="destination">
+          <th mat-header-cell *matHeaderCellDef>Destination</th>
+          <td
+            mat-cell
+            *matCellDef="let element"
+            class="col-m-w ellipsis"
+            title="{{ element.Destination }}"
+            >
+            {{ element.Destination }}
+          </td>
+        </ng-container>
+        <ng-container matColumnDef="actions">
+          <th mat-header-cell *matHeaderCellDef class="text-right">Actions</th>
+          <td mat-cell *matCellDef="let element" class="text-right">
+            <mat-icon
+              class="action-icon fill-primary"
+              svgIcon="edit_ico"
+              (click)="onEdit(element)"
+            ></mat-icon>
+          </td>
+        </ng-container>
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+      </table>
+    </div>
+  }
 </div>
