+++ src/EDAWebClient/src/app/components/settings-properties/settings-properties.component.html
@@ -5,7 +5,7 @@
       mat-raised-button
       color="primary"
       (click)="onCreate()"
-    >
+      >
       New
     </button>
     <button
@@ -13,7 +13,7 @@
       mat-raised-button
       color="warn"
       (click)="onReset()"
-    >
+      >
       Reset
     </button>
   </div>
@@ -27,57 +27,56 @@
       name="documentName"
       (ngModelChange)="onDocumentNameChange($event)"
       required
-    />
-  </mat-form-field>
+      />
+    </mat-form-field>
 
-  <div class="properties-container" *ngIf="isProcessSelected">
-    <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-      <ng-container matColumnDef="key">
-        <th mat-header-cell *matHeaderCellDef>Key</th>
-        <td
-          mat-cell
-          *matCellDef="let element"
-          class="col-m-w ellipsis"
-          title="{{ element.key }}"
-        >
-          {{ element.key }}
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="value">
-        <th mat-header-cell *matHeaderCellDef>Value</th>
-        <td
-          mat-cell
-          *matCellDef="let element"
-          class="col-m-w ellipsis"
-          title="{{ element.value }}"
-        >
-          {{ element.value }}
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="actions">
-        <th mat-header-cell *matHeaderCellDef class="text-right">Actions</th>
-        <td
-          mat-cell
-          *matCellDef="let element; let i = index"
-          class="text-right"
-        >
-          <mat-icon
-            class="action-icon fill-primary"
-            svgIcon="edit_ico"
-            (click)="onEdit(element, i)"
-          ></mat-icon>
-          <mat-icon
-            class="action-icon fill-primary"
-            svgIcon="remove_ico"
-            (click)="onDelete(element, i)"
-          ></mat-icon>
-        </td>
-      </ng-container>
-
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-    </table>
+    @if (isProcessSelected) {
+      <div class="properties-container">
+        <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
+          <ng-container matColumnDef="key">
+            <th mat-header-cell *matHeaderCellDef>Key</th>
+            <td
+              mat-cell
+              *matCellDef="let element"
+              class="col-m-w ellipsis"
+              title="{{ element.key }}"
+              >
+              {{ element.key }}
+            </td>
+          </ng-container>
+          <ng-container matColumnDef="value">
+            <th mat-header-cell *matHeaderCellDef>Value</th>
+            <td
+              mat-cell
+              *matCellDef="let element"
+              class="col-m-w ellipsis"
+              title="{{ element.value }}"
+              >
+              {{ element.value }}
+            </td>
+          </ng-container>
+          <ng-container matColumnDef="actions">
+            <th mat-header-cell *matHeaderCellDef class="text-right">Actions</th>
+            <td
+              mat-cell
+              *matCellDef="let element; let i = index"
+              class="text-right"
+              >
+              <mat-icon
+                class="action-icon fill-primary"
+                svgIcon="edit_ico"
+                (click)="onEdit(element, i)"
+              ></mat-icon>
+              <mat-icon
+                class="action-icon fill-primary"
+                svgIcon="remove_ico"
+                (click)="onDelete(element, i)"
+              ></mat-icon>
+            </td>
+          </ng-container>
+          <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+          <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+        </table>
+      </div>
+    }
   </div>
-</div>
